Updated data management. Migrated Accounts and Cards to user and implemented Feature to User on BankingScotty.

+++ src/main/java/org/poo/banking/user/account/PaymentStrategy.java
@@ -1,5 +1,5 @@
 package org.poo.banking.user.account;
 
 public interface PaymentStrategy {
-    public void pay(float amount);
+    public void pay(double amount);
 }
